organized components into folders and cleaned up GoalProgress component
