+++ components/gems/PointsBar.jsx
